kb/requirements.txt
faiss-cpu>=1.8.0
numpy>=1.26.0
sentence-transformers>=2.7.0
